Move dump action list commands to SnM_Misc
... since it's based on a REAPER's action window hack anymore (8bda5d5)
